Fix behavior of PhoneNumberToStringTransformer
The current implementation of
```PhoneNumberToStringTransformer::reverseTransform``` is incorrect:
for each given string it should either return ```null``` or an instance
of ```PhoneNumber```. If an error occurs while parsing the phone number,
the transformer should throw a ```TransformationFailedException```. This
will also automatically trigger the Form component to display a
validation error.

The fact that it currently returns the input string when it can't parse
it leads to strange behavior, especially when mapping to a property that
is mapped as a doctrine phone_number type.